requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
dnspython==2.3.0
idna==3.6
numpy==1.21.6
pandas==1.3.5
pymongo==4.6.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
urllib3==1.26.6
python-dotenv==0.21.1
